Move getnodeaddresses redefinition from v27 to v22
`getnodeaddresses` was updated in v22 not v27.

Move the redefinition to v22 and update the reexports.
