+++ types/src/v23/mod.rs
@@ -294,10 +294,10 @@ pub use crate::{
     },
     v18::{
         ActiveCommand, AnalyzePsbt, AnalyzePsbtError, AnalyzePsbtInput, AnalyzePsbtInputMissing,
-        AnalyzePsbtInputMissingError, DeriveAddresses, GetAddressInfoError, GetNodeAddresses,
+        AnalyzePsbtInputMissingError, DeriveAddresses, GetAddressInfoError,
         GetReceivedByLabel, ImportMulti, ImportMultiEntry, JoinPsbts, JsonRpcError,
         ListReceivedByLabel, ListReceivedByLabelError, ListUnspent, ListUnspentItem, ListWalletDir,
-        ListWalletDirWallet, NodeAddress, UtxoUpdatePsbt,
+        ListWalletDirWallet, UtxoUpdatePsbt,
     },
     v19::{
         Bip9SoftforkInfo, Bip9SoftforkStatistics, Bip9SoftforkStatus, GetBalances,
@@ -315,6 +315,7 @@ pub use crate::{
     v22::{
         Banned, EnumerateSigners, GetAddressInfo, GetAddressInfoEmbedded, GetMempoolInfo,
         ListBanned, ListDescriptors, MempoolAcceptance, MempoolAcceptanceError, ScriptPubkey,
-        TestMempoolAccept, TestMempoolAcceptError, WalletDisplayAddress,
+        TestMempoolAccept, TestMempoolAcceptError, WalletDisplayAddress, GetNodeAddresses,
+        NodeAddress,
     },
 };
