+++ types/src/v29/mod.rs
@@ -318,7 +318,9 @@ pub use crate::{
         AddPeerAddress, GetIndexInfo, GetIndexInfoName, ImportDescriptors, ImportDescriptorsResult,
         PsbtBumpFee, PsbtBumpFeeError, Send, SendError, UpgradeWallet,
     },
-    v22::{Banned, EnumerateSigners, ListBanned, ScriptPubkey, WalletDisplayAddress},
+    v22::{
+        Banned, EnumerateSigners, ListBanned, ScriptPubkey, WalletDisplayAddress, GetNodeAddresses, NodeAddress
+    },
     v23::{
         AddMultisigAddress, Bip9Info, Bip9Statistics, CreateMultisig, DecodeScript,
         DecodeScriptError, DeploymentInfo, GetDeploymentInfo, GetDeploymentInfoError,
@@ -344,7 +346,7 @@ pub use crate::{
         LastProcessedBlock, LastProcessedBlockError, LoadTxOutSet, LoadTxOutSetError, LoadWallet,
         PeerInfo, UnloadWallet,
     },
-    v27::{GetNodeAddresses, GetPrioritisedTransactions, NodeAddress, PrioritisedTransaction},
+    v27::{GetPrioritisedTransactions, PrioritisedTransaction},
     v28::{
         CreateWalletDescriptor, GetAddressInfo, GetAddressInfoEmbedded, GetHdKeys, GetHdKeysError,
         GetNetworkInfo, GetTransaction, HdKey, HdKeyDescriptor, Logging, SubmitPackage,
